Fix Iterator support for Inventory endpoint requests
This was a tricky architectural problem. Requests to the InventorySearch endpoint return a result for each item and location; if an item is stored in six locations, then there are six results for that item. But, I wanted application code to be able to spin through sets of items without needing to worry about de-duplication. This would require "squashing" results from InventorySearch and keeping track of how many results had been received from past requests to that endpoint (for that search), as well as dealing with situations where an item may be split across multiple InventorySearch "pages".

Would it be better to handle all of this logic in the Iterator class? What if there were other API endpoints that might need similar treatment in the future? It could get messy.

I opted to extend the support for callbacks that had already been added to the Iterator class, so that every time it needed to load a new page, it would be able to ask another class to handle the request. Then I added support for cursors to the Inventory class, and handled the squashing and incomplete item edge case there.

Limited testing has been successful so far, it seems to be working well.